Move stateful tests to public testing API
